+++ CHANGELOG.md
@@ -37,6 +37,13 @@ Line wrap the file at 100 chars.                                              Th
 - Don't hijack DNS when localhost is configured. This is more in line with other platforms.
   Unexpected DNS traffic is still blocked when leaving the host.
 - Enable IPv6 by default. This fixes DNS and routing being broken on some platforms.
+- Proxy DNS queries through a local resolver.
+
+### Fixed
+#### macOS
+- Fix Apple leak toggle not working. The issue was that DNS queries to the tunnel resolver were
+  being sent on the physical interface.
+
 
 ## [2024.6-beta1] - 2024-09-26
 ### Added